Add example launch file
This file allows to automate the starting of the rosserial node
is required to publish the sensor data to the ROS network.

That node will listen to the serial stream of data and publish
it to the sensor topic in the ROS network.

Usage:

    roslaunch src/sensor_gp2y0a02yk/getting_started.launch

See http://wiki.ros.org/roslaunch/XML